+++ lib/whois/record/parser/whois.nic.ve.rb
@@ -30,8 +30,8 @@ module Whois
         property_supported :status do
           if content_for_scanner =~ /Estatus del dominio: (.+?)\n/
             case $1.downcase
-              when "activo" then :registered
-              when "suspendido" then :suspended
+              when "activo"     then :registered
+              when "suspendido" then :inactive
               else
                 Whois.bug!(ParserError, "Unknown status `#{$1}'.")
             end